Remove MonoGame controls and add reference to MonoGame control library
This commit removes the MonoGame controls that were previously implemented directly within the MonoGame.WpfCore project. Additionally, it adds a new reference to the MonoGame.WpfCore.Control library that houses these functionalities. By extracting the controls into a separate library, the code base is better organized and allows for more reusability of the control components.